+++ src/main/java/com/mina/ml/neuralnetwork/layer/HiddenLayer.java
@@ -17,7 +17,7 @@ public class HiddenLayer extends Layer {
     }
 
     @Override
-    public float[][] forwardPropagation() {
+    public double[][] forwardPropagation() {
 //        MatrixManipulator.debugMatrix(layerName + " input:", input);
 //        MatrixManipulator.debugMatrix(layerName + " weight:", weight);
 
@@ -28,7 +28,7 @@ public class HiddenLayer extends Layer {
 //        MatrixManipulator.debugMatrix(layerName + " Z:", Z);
 
         // add column of ones in the very beginning of the matrix
-        float[][] outMatrix = MatrixManipulator.addColumnOfOnes(Z);
+        double[][] outMatrix = MatrixManipulator.addColumnOfOnes(Z);
 //        MatrixManipulator.debugMatrix(layerName + " input [*After Adding Column of Ones*]:", outMatrix);
 
         return nextLayer.input(outMatrix)
@@ -36,20 +36,25 @@ public class HiddenLayer extends Layer {
     }
 
     @Override
-    public void backPropagation(float[][] prevCostPrime) {
+    public void backPropagation(double[][] prevCostPrime) {
 //logger.info("HERE 7.1");
-        float[][] primeA = activationFunction.activatePrime(A);
+        double[][] primeA = activationFunction.activatePrime(A);
 //        MatrixManipulator.debugMatrix(layerName + " A prime: ", primeA);
+
 //logger.info("HERE 7.2");
 //        MatrixManipulator.debugMatrix(layerName + " prevCostPrime: ", prevCostPrime);
-        float[][] costOutputPrime = MatrixManipulator.multiplyEntries(prevCostPrime, primeA);
+        double[][] costOutputPrime = MatrixManipulator.multiplyEntries(prevCostPrime, primeA);
 //        MatrixManipulator.debugMatrix(layerName + " costOutputPrime: ", costOutputPrime);
+
+
 //logger.info("HERE 7.3");
         logger.debug("{} BackPropagation calculateDeltaWeight ...", layerName);
         calculateDeltaWeight(costOutputPrime);
+
 //logger.info("HERE 7.4");
         logger.debug("{} BackPropagation prepareErrorCostThenBackPropagate ...", layerName);
         prepareErrorCostThenBackPropagate(costOutputPrime);
+//        System.exit(0);
     }
 
     @Override
